Fix some other issues that I noticed while trying to fix a test regression

+++ src/buffer/conv/rgb.rs
@@ -29,7 +29,7 @@ pub fn ycc(from: &[u8; 3], to: &mut [u8; 3]) {
     *cr = ((128 * r - 107 * g - 21 * b) / 256 + 128).clamp(0, 255) as u8;
 }
 pub fn luma(from: &[u8; 3], to: &mut [u8; 1]) {
-    let [r, g, b] = from.map(|c| c as i16);
+    let [r, g, b] = from.map(|c| c as i32);
     to[0] = ((r * 77 + g * 150 + b * 29).max(0) >> 8).min(255) as u8;
 }
 pub fn gray(from: &[u8; 3], to: &mut [u8; 1]) {
